vision-microservice/requirements.txt
transformers
torch
Pillow
fastapi
uvicorn
python-multipart
requests
